test: checkpoint compilation
