+++ world-chain-builder/crates/world/node/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 [dependencies]
 world-chain-builder-payload.workspace = true
 world-chain-builder-pool.workspace = true
+world-chain-builder-pbh = { workspace = true, optional = true }
 
 reth.workspace = true
 reth-chain-state.workspace = true
@@ -28,16 +29,20 @@ reth-provider.workspace = true
 reth-prune-types.workspace = true
 reth-trie.workspace = true
 reth-trie-db.workspace = true
+reth-e2e-test-utils = { workspace = true, optional = true }
 
 alloy-eips.workspace = true
 alloy-primitives.workspace = true
 alloy-rpc-types.workspace = true
 alloy-sol-types.workspace = true
 
+revm-primitives = { workspace = true, optional = true }
+
 tokio.workspace = true
 eyre.workspace = true
 futures.workspace = true
 clap.workspace = true
+chrono = { workspace = true, optional = true }
 
 [dev-dependencies]
 world-chain-builder-pbh.workspace = true
@@ -61,4 +66,14 @@ criterion.workspace = true
 semaphore.workspace = true
 serde_json.workspace = true
 chrono.workspace = true
+
+[features]
+default = []
+test = [
+    "world-chain-builder-pool/test",
+    "world-chain-builder-pbh",
+    "dep:reth-e2e-test-utils",
+    "dep:revm-primitives",
+    "dep:chrono",
+]
