+++ world-chain-builder/crates/world/node/src/test/assets/genesis.json
